gemini/requirements.txt
flask
langchain
langchain-google-genai
google-generativeai
openai
python-dotenv
requests
